ER-453 - Renamed NextVacancyReviewProvider to NextVacancyReviewServices
