+++ src/Shared/Recruit.Vacancies.Client/Application/Services/INextVacancyReviewService.cs
@@ -1,9 +1,9 @@
 ﻿using System.Threading.Tasks;
 using Esfa.Recruit.Vacancies.Client.Domain.Entities;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Providers
+namespace Esfa.Recruit.Vacancies.Client.Application.Services
 {
-    public interface INextVacancyReviewProvider
+    public interface INextVacancyReviewService
     {
         Task<VacancyReview> GetNextVacancyReviewAsync(string userId);
     }
